pyproject.toml
[tool.poetry]
name = "nullifiercore-bot"
version = "1.2.0"
description = "Телеграм бот для выгрузки подписчиков"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "20.3"
telethon = "1.30.3"
aiohttp = "3.8.5"
python-dotenv = "1.0.0"
telegram = "^0.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"